added calculation of avgspeed and removed some old comments

+++ src/main/java/master2018/flink/VehicleTelematics.java
@@ -49,18 +49,9 @@ public class VehicleTelematics {
                         })
                 .keyBy(1);
 
-        SingleOutputStreamOperator<Tuple4<Integer, Integer, Integer, Integer>> result = keyedStream
+        SingleOutputStreamOperator<Tuple4<Integer, Integer, Integer, Double>> result = keyedStream
                 .window(EventTimeSessionWindows.withGap(Time.minutes(1)))
                 .apply(new AverageSpeedControl());
-//                .reduce(new ReduceFunction<Tuple8<Integer, Integer, Integer, Long, Integer, Boolean, Integer, Integer>>() {
-//                    @Override
-//                    public Tuple8<Integer, Integer, Integer, Long, Integer, Boolean, Integer, Integer> reduce(
-//                            Tuple8<Integer, Integer, Integer, Long, Integer, Boolean, Integer, Integer> v1,
-//                            Tuple8<Integer, Integer, Integer, Long, Integer, Boolean, Integer, Integer> v2) throws Exception {
-//                        Tuple8<Integer, Integer, Integer, Long, Integer, Boolean, Integer, Integer> out = new Tuple8<>(v1.f0, v1.f1, v1.f2, v1.f3, v1.f4, v1.f5, v1.f6, v1.f7);
-//                        return out;
-//                    }
-//                });
 //        Time1, Time2, VID, XWay, Dir, AvgSpd
         result.writeAsCsv(args[1] + "/avgspeedfines.csv", FileSystem.WriteMode.OVERWRITE);
